+++ BackgroundServices/EventBusBackgroundService.cs
@@ -1,10 +1,8 @@
-using System.Reflection;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using RabbitMqBus.DI;
 using RabbitMqBus.EventBus.Interfaces;
-using RabbitMqBus.Models;
 using RabbitMqBus.Rmq.Interfaces;
 
 namespace RabbitMqBus.BackgroundServices;
@@ -68,36 +66,9 @@ public class EventBusBackgroundService : BackgroundService
         {
             try
             {
-                MethodInfo? method;
-                
-                if (!string.IsNullOrEmpty(reg.CustomQueueName))
-                {
-                    method = typeof(IEventBus)
-                        .GetMethod(nameof(IEventBus.SubscribeAsync), 
-                            2, BindingFlags.Public | BindingFlags.Instance, types: new[] { typeof(EventExchangeType), typeof(string) })?
-                        .MakeGenericMethod(reg.EventType, reg.HandlerType);
-                    
-                    if (method != null)
-                    {
-                        await (Task)method.Invoke(eventBus, new object[] { reg.ExchangeType, reg.CustomQueueName })!;
-                        _logger.LogInformation("Подписка на {EventType} с обработчиком {HandlerType} (тип exchange: {ExchangeType}, очередь: {QueueName})",
-                            reg.EventType.Name, reg.HandlerType.Name, reg.ExchangeType, reg.CustomQueueName);
-                    }
-                }
-                else
-                {
-                    method = typeof(IEventBus)
-                        .GetMethod(nameof(IEventBus.SubscribeAsync), 
-                            2, BindingFlags.Public | BindingFlags.Instance, types: new[] { typeof(EventExchangeType) })?
-                        .MakeGenericMethod(reg.EventType, reg.HandlerType);
-                    
-                    if (method != null)
-                    {
-                        await (Task)method.Invoke(eventBus, new object[] { reg.ExchangeType })!;
-                        _logger.LogInformation("Подписка на {EventType} с обработчиком {HandlerType} (тип exchange: {ExchangeType})",
-                            reg.EventType.Name, reg.HandlerType.Name, reg.ExchangeType);
-                    }
-                }
+                await reg.SubscribeAction(eventBus);
+                _logger.LogInformation("Подписка на {EventType} с обработчиком {HandlerType}",
+                    reg.EventType.Name, reg.HandlerType.Name);
             }
             catch (Exception ex)
             {